// src/app/models/venta.model.ts
export class Venta {
    VENTA_ID: number;
    VENTA_FECHA:string;
    VENTA_SUBTOTAL: number;
    VENTA_TOTAL: number; 
    VENTA_ESTADO: number;
    COMPROBANTE_ID: number;
    USU_ID: number;
    METODO_PAGO_ID: number;
    CLIENTE_ID: number;
    USU_NOMBRES: string;
    USU_APELLIDO_PATERNO: string;
    USU_APELLIDO_MATERNO: string;
    CLIENTE_NOMBRES: string;
    CLIENTE_APELLIDOS: string;
    COMPROBANTE_TIPO: string;   
    METODO_PAGO_DESCRIPCION: string;
}